Backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
pandas
numpy